Move code out to a separate package
