+++ MSBuild.Sdk.SqlProj.sln
@@ -20,6 +20,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DacpacTool.Tests", "test\Da
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestAspireHost", "test\TestAspireHost\TestAspireHost.csproj", "{6691FBBD-479C-4590-8964-DD08909287EB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSBuild.Sdk.SqlProj.Aspire", "src\MSBuild.Sdk.SqlProj.Aspire\MSBuild.Sdk.SqlProj.Aspire.csproj", "{78924A88-997D-4D20-BDED-7315745C7C5F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -78,6 +80,18 @@ Global
 		{6691FBBD-479C-4590-8964-DD08909287EB}.Release|x64.Build.0 = Release|Any CPU
 		{6691FBBD-479C-4590-8964-DD08909287EB}.Release|x86.ActiveCfg = Release|Any CPU
 		{6691FBBD-479C-4590-8964-DD08909287EB}.Release|x86.Build.0 = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|x64.Build.0 = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Debug|x86.Build.0 = Debug|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|x64.ActiveCfg = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|x64.Build.0 = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|x86.ActiveCfg = Release|Any CPU
+		{78924A88-997D-4D20-BDED-7315745C7C5F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -87,6 +101,7 @@ Global
 		{78CA944A-928F-48FA-B29F-C5DC96E67684} = {B0420F9B-A902-4E9D-8B29-55B4F626483B}
 		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4} = {A3A4D55B-4484-4501-894F-0B07708A9A1D}
 		{6691FBBD-479C-4590-8964-DD08909287EB} = {A3A4D55B-4484-4501-894F-0B07708A9A1D}
+		{78924A88-997D-4D20-BDED-7315745C7C5F} = {B0420F9B-A902-4E9D-8B29-55B4F626483B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {832DCF12-9633-4AEB-A9D8-444EDF649047}
