+++ src/MSBuild.Sdk.SqlProj.Aspire/TargetDatabaseResourceAnnotation.cs
@@ -1,0 +1,7 @@
+using Aspire.Hosting.ApplicationModel;
+
+namespace MSBuild.Sdk.SqlProj.Aspire;
+
+public record TargetDatabaseResourceAnnotation(string TargetDatabaseResourceName) : IResourceAnnotation
+{
+}
