+++ test/TestAspireHost/Program.cs
@@ -3,7 +3,7 @@ var builder = DistributedApplication.CreateBuilder(args);
 var sql = builder.AddSqlServer("sql")
                  .AddDatabase("test");
 
-builder.AddDatabaseProject<Projects.TestProject>("db")
+builder.AddDataTierApplication<Projects.TestProject>("db")
        .PublishTo(sql);
 
 builder.Build().Run();
